+++ framework/containers/array_byte.h
@@ -17,7 +17,7 @@ void array_byte_resize(struct Array_Byte * array, uint64_t target_capacity);
 void array_byte_push(struct Array_Byte * array, uint8_t value);
 void array_byte_push_many(struct Array_Byte * array, uint64_t count, uint8_t const * value);
 
-void array_byte_set_many(struct Array_Byte * array, uint32_t index, uint32_t count, uint8_t const * value);
+void array_byte_set_many(struct Array_Byte * array, uint64_t index, uint64_t count, uint8_t const * value);
 
 uint8_t array_byte_pop(struct Array_Byte * array);
 uint8_t array_byte_peek(struct Array_Byte const * array, uint64_t offset);
